requirement.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-jose
python-decouple
